refactor: renaming hydraauth to hydra
