+++ src/index.ts
@@ -3,7 +3,7 @@ import express, { Request, NextFunction, Response } from 'express'
 import handlebars from 'express-handlebars'
 import request from 'request'
 import { authHandler } from './routes/auth'
-import hydraauth from './routes/hydraauth'
+import hydraauth from './routes/hydra'
 import { getConsent, postConsent } from './routes/consent'
 import errorHandler from './routes/error'
 import dashboard from './routes/dashboard'
